+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -5,8 +5,12 @@
 * LogExporter to correctly map Severity to OTLP.
   ([#3177](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3177))
 
-* LogExporter to special case {OriginalFormat} to populate
+* LogExporter to special case `{OriginalFormat}` to populate
   Body. ([#3182](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3182))
+
+* LogExporter to populate attributes from `LogRecord.State`, if it
+  is of type `IReadOnlyList<KeyValuePair<string, object>>`
+  ([#3182](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3182))
 
 ## 1.2.0-rc5